requirements.txt
pandas
numpy
tweepy
python-dotenv
selenium
bs4
ipykernel